+++ tests/conftest.py
@@ -34,4 +34,3 @@ class Helpers:
 @pytest.fixture
 def helpers():
     return Helpers
-