+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## 0.7.0 (XX YYY 2020)
 
 * Python 2 support dropped
+* RosePlot added
 
 ### 0.6.3 (6 Dec 2019)